fix linter and create persistent_queue_context
